backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.15.0",
    "gspread>=6.2.1",
    "oauth2client>=4.1.3",
    "python-dotenv>=1.1.0",
    "selenium>=4.32.0",
]
